Trying to pinpoint where the circular references come into play

+++ kernel/src/include/mem/pmm.h
@@ -43,9 +43,9 @@ struct contiguous_page_range {
 
 struct buddy_block {
     struct buddy_block *next;
-    uint64 zone;
+    uint8 zone;
     uint64 start_address;
-    uint64 flags;
-    uint64 order;
-    uint64 is_free;
+    uint8 flags;
+    uint8 order;
+    uint8 is_free;
 };
